Added database functioning available for items

+++ src/main/java/DTO/ItemDTO.java
@@ -5,7 +5,7 @@ public class ItemDTO implements DTO {
     public final double price;
     public final long machineCode;
 
-    public ItemDTO(int machineCode, String name, double price) {
+    public ItemDTO(long machineCode, String name, double price) {
         this.name = name;
         this.price = price;
         this.machineCode = machineCode;
